`Docs` Move responsive padding prop implementation from Card to Box
