+++ polaris.shopify.com/src/components/Box/index.tsx
@@ -1,10 +1,21 @@
 import {forwardRef} from 'react';
-import {className as classNames} from '../../utils/various';
+import {
+  className as classNames,
+  getResponsiveProps,
+  ResponsiveProp,
+} from '../../utils/various';
 import type {ClassName} from '../../utils/various';
 import type * as Polymorphic from '@radix-ui/react-polymorphic';
+import {SpaceScale} from '@shopify/polaris-tokens';
+import styles from './Box.module.scss';
 
 export interface BoxProps {
   className?: ClassName;
+  padding?: ResponsiveProp<SpaceScale>;
+  paddingInlineStart?: ResponsiveProp<SpaceScale>;
+  paddingInlineEnd?: ResponsiveProp<SpaceScale>;
+  paddingBlockStart?: ResponsiveProp<SpaceScale>;
+  paddingBlockEnd?: ResponsiveProp<SpaceScale>;
 }
 
 export type OwnProps<T> = Polymorphic.OwnProps<T>;
@@ -16,17 +27,22 @@ type PolymorphicBox = Polymorphic.ForwardRefComponent<'div', BoxProps>;
  * built. It renders a `div` element by default, customisable via the `as` prop.
  */
 export const Box = forwardRef(
-  ({as: Tag = 'div', className, ...props}, forwardedRef) => (
-    <Tag
-      {...(className
-        ? {
-            className: classNames(className),
-          }
-        : null)}
-      {...props}
-      ref={forwardedRef}
-    />
-  ),
+  ({as: Tag = 'div', className, padding = '0', ...props}, forwardedRef) => {
+    const style = {
+      ...getResponsiveProps('card', 'padding-block-start', 'space', padding),
+      ...getResponsiveProps('card', 'padding-block-end', 'space', padding),
+      ...getResponsiveProps('card', 'padding-inline-start', 'space', padding),
+      ...getResponsiveProps('card', 'padding-inline-end', 'space', padding),
+    };
+    return (
+      <Tag
+        style={style}
+        className={classNames(styles.Box, className)}
+        {...props}
+        ref={forwardedRef}
+      />
+    );
+  },
 ) as PolymorphicBox;
 
 Box.displayName = 'Box';
