+++ polaris.shopify.com/src/components/Card/index.tsx
@@ -1,5 +1,5 @@
 import styles from './Card.module.scss';
-import {getResponsiveProps, ResponsiveProp} from '../../utils/various';
+import {ResponsiveProp} from '../../utils/various';
 import {Box} from '../Box';
 import React from 'react';
 import {SpaceScale} from '@shopify/polaris-tokens';
@@ -12,14 +12,8 @@ export function Card({
   padding?: ResponsiveProp<SpaceScale>;
   className?: string;
 }>) {
-  const style = {
-    ...getResponsiveProps('card', 'padding-block-start', 'space', padding),
-    ...getResponsiveProps('card', 'padding-block-end', 'space', padding),
-    ...getResponsiveProps('card', 'padding-inline-start', 'space', padding),
-    ...getResponsiveProps('card', 'padding-inline-end', 'space', padding),
-  };
   return (
-    <Box style={style} className={[styles.Card, className]}>
+    <Box padding={padding} className={[styles.Card, className]}>
       {children}
     </Box>
   );
